Extract DB merge tests from production module
